更新 FastNoiseLite.h 以修复 MSVC 警告 C4251 完善 FastNoiseLite.h 中的类型转换 SDK功能完善

+++ sdk/plugins/TerrainOverhaul/include/FastNoiseLite.h
@@ -28,8 +28,8 @@ public:
         
         // Very basic pseudo-noise for demonstration purposes
         // Real implementation would use lookups and gradient vectors
-        return (sin(x * 12.9898 + mSeed) * 43758.5453 
-              + cos(y * 78.233 + z * 32.1) * 21321.2) - floor((sin(x) + cos(y))*10000);
+        return static_cast<float>((sin(x * 12.9898 + mSeed) * 43758.5453 
+              + cos(y * 78.233 + z * 32.1) * 21321.2) - floor((sin(x) + cos(y))*10000));
     }
     
     // Fractal Brownian Motion
